+++ src/tmtccmd/com/dummy.py
@@ -4,9 +4,8 @@ external hardware or an extra socket
 
 from __future__ import annotations
 
-from typing import Optional
+from typing import Any, Optional
 
-from deprecated.sphinx import deprecated
 from spacepackets.ccsds.time import CdsShortTimestamp
 from spacepackets.ecss.pus_1_verification import (
     RequestId,
@@ -16,7 +15,7 @@ from spacepackets.ecss.pus_1_verification import (
 from spacepackets.ecss.pus_17_test import Service17Tm
 from spacepackets.ecss.tc import PusTelecommand
 
-from tmtccmd.com import ComInterface
+from com_interface import ComInterface
 from tmtccmd.config import CoreComInterfaces
 from tmtccmd.pus.s1_verification import Subservice as Pus1Subservice
 from tmtccmd.pus.s17_test import Subservice as Pus17Subservice
@@ -30,14 +29,7 @@ class DummyHandler:
         self.current_ssc = 0
         self.reply_pending = False
 
-    @deprecated(
-        version="6.0.0",
-        reason="Use insert_telecommand instead",
-    )
-    def pass_telecommand(self, data: bytearray):
-        self.insert_telecommand(data)
-
-    def insert_telecommand(self, data: bytes):
+    def insert_telecommand(self, data: bytearray | bytes):
         self.last_tc = PusTelecommand.unpack(data)
         self.reply_pending = True
         self.generate_reply_package()
@@ -107,7 +99,7 @@ class DummyHandler:
             return []
 
 
-class DummyComIF(ComInterface):
+class DummyInterface(ComInterface):
     def __init__(self):
         self.com_if_id = CoreComInterfaces.DUMMY.value
         self.dummy_handler = DummyHandler()
@@ -118,25 +110,42 @@ class DummyComIF(ComInterface):
     def id(self) -> str:
         return self.com_if_id
 
-    def initialize(self, args: any = None) -> any:
+    def initialize(self, args: Any = None) -> None:
         self.initialized = True
 
-    def open(self, args: any = None) -> None:
+    def open(self, args: Any = None) -> None:
         self._open = True
 
     def is_open(self) -> bool:
         return self._open
 
-    def close(self, args: any = None) -> None:
+    def close(self, args: Any = None) -> None:
         self._open = False
 
-    def data_available(self, timeout: float = 0, parameters: any = 0):
+    def receive(self, parameters: Any = 0) -> list[bytes]:
+        """Returns a list of received packets. The child class can use a separate thread to poll for
+        the packets or use some other mechanism and container like a deque to store packets
+        to be returned here.
+
+        :param parameters:
+        :raises ReceptionDecodeError: If the underlying COM interface uses encoding and
+            decoding and the decoding fails, this exception will be returned.
+        :return:
+        """
+        return self.dummy_handler.receive_reply_package()
+
+    def packets_available(self, parameters: Any = 0) -> int:
+        """Poll whether packets are available.
+
+        :param parameters: Can be an arbitrary parameter.
+        :raises ReceptionDecodeError: If the underlying COM interface uses encoding and
+            decoding when determining the number of available packets, this exception can be
+            thrown on decoding errors.
+        :return: Number of packets available.
+        """
         if self.dummy_handler.reply_pending:
             return True
         return False
-
-    def receive(self, parameters: any = 0) -> list[bytes]:
-        return self.dummy_handler.receive_reply_package()
 
     def send(self, data: bytes | bytearray):
         if data is not None: