+++ src/tmtccmd/logging/__init__.py
@@ -28,7 +28,7 @@ def add_colorlog_console_logger(logger: logging.Logger, log_level: int = logging
     )
     custom_formatter = CustomTmtccmdFormatter(
         info_fmt=(
-            "%(log_color)s%(levelname)-8s %(cyan)s%(asctime)s." "%(msecs)03d %(reset)s%(message)s"
+            "%(log_color)s%(levelname)-8s %(cyan)s%(asctime)s.%(msecs)03d %(reset)s%(message)s"
         ),
         dbg_fmt=dbg_fmt,
         err_fmt=dbg_fmt,
@@ -44,7 +44,7 @@ def add_colorlog_console_logger(logger: logging.Logger, log_level: int = logging
 
 def add_error_file_logger(logger: logging.Logger):
     file_format = logging.Formatter(
-        fmt=("%(levelname)-8s: %(asctime)s.%(msecs)03d [%(filename)s:%(lineno)d]" " %(message)s"),
+        fmt=("%(levelname)-8s: %(asctime)s.%(msecs)03d [%(filename)s:%(lineno)d] %(message)s"),
         datefmt="%Y-%m-%d %H:%M:%S",
     )
     # TODO: Use path relative to script dir, otherwise this craps everything