+++ src/tmtccmd/tmtc/handler.py
@@ -1,6 +1,6 @@
 from abc import abstractmethod, ABC
 
-from tmtccmd.com import ComInterface
+from com_interface import ComInterface
 from tmtccmd.tmtc.procedure import ProcedureWrapper
 from tmtccmd.tmtc.queue import QueueWrapper, QueueEntryHelper